Componentise IIIFViewer
